+++ redwood-protocol-guest/src/commonTest/kotlin/app/cash/redwood/protocol/guest/GeneratedProtocolBridgeTest.kt
@@ -29,8 +29,8 @@ import app.cash.redwood.protocol.WidgetTag
 import assertk.assertThat
 import assertk.assertions.hasMessage
 import assertk.assertions.isEqualTo
-import com.example.redwood.testing.compose.TestSchemaProtocolBridge
 import com.example.redwood.testing.compose.TestScope
+import com.example.redwood.testing.protocol.guest.TestSchemaProtocolBridge
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.time.Duration